feat: add PHPDoc property annotations to all models, GitHub CI workflows, and delegation tests
- Add @property PHPDoc blocks to all Eloquent models for IDE type safety and static analysis
- Add @phpstan-ignore suppression comments across services, jobs, and controllers
- Add GitHub Actions CI workflows: phpstan, pint, test, vitest, eslint, composer-audit, npm-audit, build
- Add TrustGatedApprovalTest feature test for Delegation module
- Update storage/.gitignore and storage/framework/.gitignore
- Add config/agent.php additions
- Update OpenAPI spec (docs/openapi/agent-api-v1.json)

+++ app/Console/Commands/Memory/MemoryStatsCommand.php
@@ -257,10 +257,10 @@ class MemoryStatsCommand extends Command
             ['Provider', 'Requests', 'Input Tokens', 'Output Tokens', 'Est. Cost (USD)'],
             $usage->map(fn ($u) => [
                 $u->provider,
-                number_format($u->requests),
-                number_format($u->input_tokens ?? 0),
-                number_format($u->output_tokens ?? 0),
-                '$'.number_format($u->cost ?? 0, 4),
+                number_format((int) $u->requests),
+                number_format((int) ($u->input_tokens ?? 0)),
+                number_format((int) ($u->output_tokens ?? 0)),
+                '$'.number_format((float) ($u->cost ?? 0), 4),
             ])->toArray()
         );
     }
